Add shell script to automate building of packages for release (#5)
Update version in pubspec to match first  GH release. 
This is largely symbolic because the build aar and build ios-framework
commands ignore this version.

Add shell script to automate building of packages for release
There's some custom flags that you want to use to make the output
less confusing for developers. This gives us an easy 1 invoke command to
do all that for us.